Renamed model.go to options.go and moved it in pad_options folder in model directory
